:recycle: [BE_MY] Update VisualTagValue methods

+++ backend/src/main/java/com/pjt3/promise/response/UsersTagGetRes.java
@@ -5,5 +5,6 @@ import lombok.Data;
 @Data
 public class UsersTagGetRes {
 	String tagName;
-	int tagValue;
+//	int tagValue;
+	double tagValue;
 }
